+++ R/campaigns.R
@@ -6,7 +6,7 @@
 #' @source 84.51, Customer Journey study, \url{http://www.8451.com/area51/}
 #' @format A data frame with 6,589 rows and 2 variables
 #' \describe{
-#' \item{campaign_id}{Uniquely identifies each campaign; Ranges 1-30}
+#' \item{campaign_id}{Uniquely identifies each campaign; Ranges 1-27}
 #' \item{household_id}{Uniquely identifies each household}
 #' }
 #' @examples